requirements.txt
cmcrameri==1.7
h5py==3.10.0
matplotlib==3.8.0
numpy==1.26.0
pandas==2.2.3
pysolar==0.11
python-dotenv==1.0.1
PyYAML==6.0.2
scipy==1.14.1
seaborn==0.13.2
statsmodels==0.14.0
